Move blockToNativeValue from SPI predicate.Utils to type.TypeUtils
It's a better place, it's where other utils like this live.
