+++ core/trino-spi/pom.xml
@@ -228,6 +228,11 @@
                                 </item>
                                 <item>
                                     <ignore>true</ignore>
+                                    <code>java.method.removed</code>
+                                    <old>method java.lang.Object io.trino.spi.predicate.Utils::blockToNativeValue(io.trino.spi.type.Type, io.trino.spi.block.Block)</old>
+                                </item>
+                                <item>
+                                    <ignore>true</ignore>
                                     <code>java.method.noLongerDefault</code>
                                     <old>method io.trino.spi.connector.ConnectorPageSink io.trino.spi.connector.ConnectorPageSinkProvider::createPageSink(io.trino.spi.connector.ConnectorTransactionHandle, io.trino.spi.connector.ConnectorSession, io.trino.spi.connector.ConnectorInsertTableHandle, java.util.Optional&lt;io.trino.spi.connector.ConnectorTableCredentials&gt;, io.trino.spi.connector.ConnectorPageSinkId)</old>
                                     <justification>Removed deprecated method without ConnectorTableCredentials parameter</justification>
