+++ core/trino-spi/src/main/java/io/trino/spi/predicate/EquatableValueSet.java
@@ -41,6 +41,7 @@ import static io.trino.spi.function.InvocationConvention.InvocationReturnConvent
 import static io.trino.spi.function.InvocationConvention.simpleConvention;
 import static io.trino.spi.predicate.Utils.TUPLE_DOMAIN_TYPE_OPERATORS;
 import static io.trino.spi.predicate.Utils.handleThrowable;
+import static io.trino.spi.type.TypeUtils.blockToNativeValue;
 import static io.trino.spi.type.TypeUtils.writeNativeValue;
 import static java.lang.String.format;
 import static java.util.Collections.unmodifiableSet;
@@ -468,7 +469,7 @@ public class EquatableValueSet
 
         public Object getValue()
         {
-            return Utils.blockToNativeValue(type, block);
+            return blockToNativeValue(type, block);
         }
 
         @Override
